+++ app/src/main/java/com/android/collections/models/product_detalis/Voting.kt
@@ -5,4 +5,4 @@ import com.google.gson.annotations.SerializedName
 data class Voting(@SerializedName("voting_count")
                   var votingCount: Int? = null,
                   @SerializedName("all_voting")
-                  var allVoting: Int? = null)
+                  var allVoting: Double? = null)
